requirements.txt
requests
python-dotenv
newsapi-python
datetime
pandas
pytest
